split conversation adapter into separate class
